requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiosmtplib==3.0.1
aiofiles==23.2.1
httpx==0.25.2
pillow==10.1.0
jinja2==3.1.2
email-validator==2.1.0